requirements.txt
PyOpenGL>=3.1.7
glfw>=2.7.0
numpy>=1.26.0
Pillow>=10.0.0
PyYAML>=6.0.1

